requirements.txt
streamlit>=1.28.0
langchain-openai>=0.1.0
python-dotenv>=1.0.0
pandas>=2.0.0
msal>=1.24.0
requests>=2.31.0
flask>=2.0.0
playwright>=1.40.0
pillow>=10.0.0
